Upgrade the colormath plugin
The latest colormath plugin requires new versions of just about everything, so the other dependencies were upgraded as well.

This brought with it some rolling dependencies, so currently qodana is (mostly) non-functional.

The plugin builds and seems to work, though. Still, marking this as WIP until we can get some other eyes on this.