Fixed: null with missing lamdera compiler
This fixes the various paths for the settings to ensure that the paths exist before setting them.

It also contains a few automated fixes suggested by the IDE to upgrade some parts to the latest methods. Specifically, an error was being raised on doing `project.save()` within a write action. ChatGPT suggested moving the `project.save()` call outside of the write action, which seems to have fixed it.

Updated the github repo in the plugin.xml file as well.

Should fix #3.